Add files via upload
A telephone and internet service company wants to detect clients that could drop out its services to offer a discont as deterrence. That is of paramount importance, given that not offering a discount to someone who is planning to leave implies a cost as much as offering to someone who would not have left. The company suggest the AUC-ROC metric for the evaluation of the model.